+++ CoreLibrary/pipe.h
@@ -44,7 +44,7 @@ namespace core {
 // N1: N writers, 1 reader
 // NN: N writers, N readers
 #ifdef PIPE_1
-template<typename T, uint32 _S> class Pipe11:
+template<typename T, uint64 _S> class Pipe11:
     public Semaphore,
     public CriticalSection {
 private:
@@ -59,8 +59,8 @@ private:
             if (next) delete next;
         }
     };
-    int32 head;
-    int32 tail;
+    int64 head;
+    int64 tail;
     Block *first;
     Block *last;
     Block *spare;
@@ -75,7 +75,7 @@ public:
     T pop(); // decreases the size as necessary
 };
 
-template<typename T, uint32 _S> class Pipe1N:
+template<typename T, uint64 _S> class Pipe1N:
     public Pipe11<T, _S> {
 private:
     CriticalSection popCS;
@@ -86,7 +86,7 @@ public:
     T pop();
 };
 
-template<typename T, uint32 _S> class PipeN1:
+template<typename T, uint64 _S> class PipeN1:
     public Pipe11<T, _S> {
 private:
     CriticalSection pushCS;
@@ -97,7 +97,7 @@ public:
     void push(T &t);
 };
 
-template<typename T, uint32 _S> class PipeNN:
+template<typename T, uint64 _S> class PipeNN:
     public Pipe11<T, _S> {
 private:
     CriticalSection pushCS;
@@ -110,25 +110,25 @@ public:
     T pop();
 };
 #elif defined PIPE_2
-template<typename T, uint32 _S, class Pipe> class Push1;
-template<typename T, uint32 _S, class Pipe> class PushN;
-template<typename T, uint32 _S, class Pipe> class Pop1;
-template<typename T, uint32 _S, class Pipe> class PopN;
+template<typename T, uint64 _S, class Pipe> class Push1;
+template<typename T, uint64 _S, class Pipe> class PushN;
+template<typename T, uint64 _S, class Pipe> class Pop1;
+template<typename T, uint64 _S, class Pipe> class PopN;
 
 // a Pipe<T,_S> is a linked list of blocks containing _S objects of type T
 // push() adds an object at the tail of the last block and moves the tail forward; when the tail reaches the end of the last block, a new block is appended to the list
 // pop() moves the head forward and returns the object at this location; when the head reaches the end of the first block, this block is deallocated
 // Synchronization between readers and writers is lock-free under no contention (reqCount), and uses a lock (Semaphore::) under contention
-// single writer pipes can use an int32 tail, whereas multiple writer versions require int32 volatile tail; idem for readers
+// single writer pipes can use an int64 tail, whereas multiple writer versions require int64 volatile tail; idem for readers
 // The Head and Tail argument is meant to allow the parameterizing of heads and tails
 // Push and Pop are functors tailored to the multiplicity of resp. the read and write threads
 // P is the actual pipe class (e.g. Pipe11, etc.)
-template<typename T, uint32 _S, typename Head, typename Tail, class P, template<typename, uint32, class> class Push, template<typename, uint32, class> class Pop> class Pipe:
+template<typename T, uint64 _S, typename Head, typename Tail, class P, template<typename, uint64, class> class Push, template<typename, uint64, class> class Pop> class Pipe:
     public Semaphore {
-    template<typename T, uint32 _S, class Pipe> friend class Push1;
-    template<typename T, uint32 _S, class Pipe> friend class PushN;
-    template<typename T, uint32 _S, class Pipe> friend class Pop1;
-    template<typename T, uint32 _S, class Pipe> friend class PopN;
+    template<typename T, uint64 _S, class Pipe> friend class Push1;
+    template<typename T, uint64 _S, class Pipe> friend class PushN;
+    template<typename T, uint64 _S, class Pipe> friend class Pop1;
+    template<typename T, uint64 _S, class Pipe> friend class PopN;
 protected:
     class Block {
     public:
@@ -147,7 +147,7 @@ protected:
 
     Head head; // starts at -1
     Tail tail; // starts at 0
-    int32 volatile waitingList; // amount of readers that have to wait, negative value indicate free lunch
+    int64 volatile waitingList; // amount of readers that have to wait, negative value indicate free lunch
 
     Push<T, _S, P> *_push;
     Pop<T, _S, P> *_pop;
@@ -168,14 +168,14 @@ protected:
     PipeFunctor(Pipe &p);
 };
 
-template<typename T, uint32 _S, class Pipe> class Push1:
+template<typename T, uint64 _S, class Pipe> class Push1:
     public PipeFunctor<Pipe> {
 public:
     Push1(Pipe &p);
     void operator()(T &t);
 };
 
-template<typename T, uint32 _S, class Pipe> class PushN:
+template<typename T, uint64 _S, class Pipe> class PushN:
     public PipeFunctor<Pipe>,
     public Semaphore {
 public:
@@ -183,14 +183,14 @@ public:
     void operator()(T &t);
 };
 
-template<typename T, uint32 _S, class Pipe> class Pop1:
+template<typename T, uint64 _S, class Pipe> class Pop1:
     public PipeFunctor<Pipe> {
 public:
     Pop1(Pipe &p);
     T operator()();
 };
 
-template<typename T, uint32 _S, class Pipe> class PopN:
+template<typename T, uint64 _S, class Pipe> class PopN:
     public PipeFunctor<Pipe>,
     public Semaphore {
 public:
@@ -198,29 +198,29 @@ public:
     T operator()();
 };
 
-template<typename T, uint32 S> class Pipe11:
-    public Pipe<T, S, int32, int32, Pipe11<T, S>, Push1, Pop1> {
+template<typename T, uint64 S> class Pipe11:
+    public Pipe<T, S, int64, int64, Pipe11<T, S>, Push1, Pop1> {
 public:
     Pipe11();
     ~Pipe11();
 };
 
-template<typename T, uint32 S> class Pipe1N:
-    public Pipe<T, S, int32, int32 volatile, Pipe1N<T, S>, Push1, PopN> {
+template<typename T, uint64 S> class Pipe1N:
+    public Pipe<T, S, int64, int64 volatile, Pipe1N<T, S>, Push1, PopN> {
 public:
     Pipe1N();
     ~Pipe1N();
 };
 
-template<typename T, uint32 S> class PipeN1:
-    public Pipe<T, S, int32 volatile, int32, PipeN1<T, S>, PushN, Pop1> {
+template<typename T, uint64 S> class PipeN1:
+    public Pipe<T, S, int64 volatile, int64, PipeN1<T, S>, PushN, Pop1> {
 public:
     PipeN1();
     ~PipeN1();
 };
 
-template<typename T, uint32 S> class PipeNN:
-    public Pipe<T, S, int32 volatile, int32 volatile, PipeNN<T, S>, PushN, PopN> {
+template<typename T, uint64 S> class PipeNN:
+    public Pipe<T, S, int64 volatile, int64 volatile, PipeNN<T, S>, PushN, PopN> {
 public:
     PipeNN();
     ~PipeNN();
